requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
nltk
wordcloud
pillow
plotly